05-Token Refreshing and Flask-JWT Extended/code/requirements.txt
Flask-JWT
Flask-RESTful
Flask-SQLAlchemy
